// src/interfaces/ScheduleItemNormalized.ts
export interface ScheduleItemNormalized {
  audience: string;
  groupName: string;
  date: string;
  dayName: string;
  dayNumber: number;
  email: string;
  lessonName: string;
  code: number;
  start: string;
  end: string;
  subGroupNumber: number;
  teacher: string;
  weekType: number;
  semesterCode: number;
  weekStart: number;
  weekEnd: number;
}